+++ lib/lightning_web/live/job_live/credential_picker.ex
@@ -64,11 +64,12 @@ defmodule LightningWeb.JobLive.CredentialPicker do
         value={@form[:keychain_credential_id].value}
       />
 
-      <div :if={!@disabled} class="text-right">
+      <div class="text-right">
         <button
           id="new-credential-button"
           type="button"
           class="link text-xs"
+          disabled={@disabled}
           phx-click="toggle_job_credential_modal"
         >
           New credential